requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
requests==2.31.0
google-api-python-client==2.100.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
pandas==2.1.1
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
openpyxl==3.1.2
python-dotenv==1.0.0
Werkzeug==2.3.7

